requirements.txt
requests=2.19.1
tensorflow_gpu=1.8.0
tqdm=4.23.4
scipy=1.1.0
